config-service/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
psycopg2-binary>=2.9.1
pydantic>=1.8.2
alembic>=1.7.1
python-dotenv>=0.19.0
pytest>=6.2.5
httpx>=0.19.0
python-multipart>=0.0.5
